Seal Pattern<T> hierarchy
